bpmn/bpmn-gv/requirements.txt
# common
argparse
colorama
httplib2
lxml
numpy
oauth2client
pandas
PyDrive
pygsheets
PyYAML
termcolor
colorama
yamllint
pysvg
pysvg-py3
svg.path